requirements.txt
pandas==2.3.0
openpyxl==3.1.5
python-pptx==1.0.2
